refactor(Answer): Answer를 뷰영역으로 이동하고 잘못된 입력시 재입력 받을 수 있도록 컨트롤러 로직 수정
