+++ src/test/kotlin/view/AnswerTest.kt
@@ -1,4 +1,4 @@
-package domain
+package view
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
